enh: issue#13 refactoring code
