+++ tests/test_imports.py
@@ -3,20 +3,20 @@ import pytest
 
 def test_import_data_ingestion():
     try:
-        import house_price_prediction.data_ingestion
+        from house_price_prediction.data_ingestion_package import data_ingestion
     except ImportError:
         pytest.fail("Failed to import data_ingestion")
 
 
 def test_import_training():
     try:
-        import house_price_prediction.training
+        from house_price_prediction.training_package import training
     except ImportError:
         pytest.fail("Failed to import training")
 
 
 def test_import_scoring():
     try:
-        import house_price_prediction.scoring
+        from house_price_prediction.scoring_package import scoring
     except ImportError:
         pytest.fail("Failed to import scoring")
